--- src/components/layout/main/MainElement.js
@@ -1,13 +1,0 @@
-import styled from 'styled-components'
-
-export const Main = styled.div`
-  font-family: 'Nunito', Helvetica, Arial, Lucida, sans-serif;
-  font-size: 0.95rem;
-  line-height: 1.5;
-  font-weight: 500;
-  overflow: hidden;
-  flex: 1;
-  position: relative;
-  text-align: justify;
-  color: #771700;
-`
