--- src/views/error/404/404Elements.js
@@ -1,17 +1,0 @@
-import { Form } from 'react-bootstrap'
-import styled from 'styled-components'
-
-export const SForm = styled(Form)`
-  background-color: #fafafa;
-  box-shadow: 0px 2px 15px 6px rgba(0, 0, 0, 0.11);
-  padding: 100px;
-  margin: 115px 0px;
-  position: center;
-`
-export const STextForm = styled.h2`
-  line-height: 1.5;
-  font-weight: 500;
-  padding-left: 10px;
-  border-left: 1px solid #771700;
-  margin: 30px 0px 50px 0px;
-  `
